backend/requirements.txt
python-dotenv
qdrant-client
sentence-transformers
uvicorn
fastapi